+++ codegen-server/src/main/kotlin/software/amazon/smithy/rust/codegen/server/smithy/RustCodegenServerPlugin.kt
@@ -73,8 +73,12 @@ class RustCodegenServerPlugin : SmithyBuildPlugin {
                 .let { StreamingShapeSymbolProvider(it, model) }
                 // Add Rust attributes (like `#[derive(PartialEq)]`) to generated shapes
                 .let { BaseSymbolMetadataProvider(it, model, additionalAttributes = listOf()) }
-                // Streaming shapes need different derives (e.g. they cannot derive Eq)
+                // Constrained shapes generate newtypes that need the same derives we place on types generated from aggregate shapes.
+                .let { ConstrainedShapeSymbolMetadataProvider(it, model, constrainedTypes) }
+                // Streaming shapes need different derives (e.g. they cannot derive `PartialEq`)
                 .let { StreamingShapeMetadataProvider(it, model) }
+                // Derive `Eq` and `Hash` if possible.
+                .let { DeriveEqAndHashSymbolMetadataProvider(it, model) }
                 // Rename shapes that clash with Rust reserved words & and other SDK specific features e.g. `send()` cannot
                 // be the name of an operation input
                 .let { RustReservedWordSymbolProvider(it, model) }