+++ rust-runtime/aws-smithy-types/src/lib.rs
@@ -14,112 +14,21 @@
 )]
 
 use crate::error::{TryFromNumberError, TryFromNumberErrorKind};
-use std::collections::HashMap;
+
+mod blob;
+pub use blob::Blob;
 
 pub mod base64;
 pub mod date_time;
+mod document;
 pub mod endpoint;
 pub mod error;
 pub mod primitive;
 pub mod retry;
 pub mod timeout;
-
-pub use crate::date_time::DateTime;
+pub use date_time::DateTime;
+pub use document::Document;
 pub use error::Error;
-
-/// Binary Blob Type
-///
-/// Blobs represent protocol-agnostic binary content.
-#[derive(Debug, PartialEq, Clone)]
-pub struct Blob {
-    inner: Vec<u8>,
-}
-
-impl Blob {
-    /// Creates a new blob from the given `input`.
-    pub fn new<T: Into<Vec<u8>>>(input: T) -> Self {
-        Blob {
-            inner: input.into(),
-        }
-    }
-
-    /// Consumes the `Blob` and returns a `Vec<u8>` with its contents.
-    pub fn into_inner(self) -> Vec<u8> {
-        self.inner
-    }
-}
-
-impl AsRef<[u8]> for Blob {
-    fn as_ref(&self) -> &[u8] {
-        &self.inner
-    }
-}
-
-/* ANCHOR: document */
-
-/// Document Type
-///
-/// Document types represents protocol-agnostic open content that is accessed like JSON data.
-/// Open content is useful for modeling unstructured data that has no schema, data that can't be
-/// modeled using rigid types, or data that has a schema that evolves outside of the purview of a model.
-/// The serialization format of a document is an implementation detail of a protocol.
-#[derive(Debug, Clone, PartialEq)]
-pub enum Document {
-    /// JSON object
-    Object(HashMap<String, Document>),
-    /// JSON array
-    Array(Vec<Document>),
-    /// JSON number
-    Number(Number),
-    /// JSON string
-    String(String),
-    /// JSON boolean
-    Bool(bool),
-    /// JSON null
-    Null,
-}
-
-impl From<bool> for Document {
-    fn from(value: bool) -> Self {
-        Document::Bool(value)
-    }
-}
-
-impl From<String> for Document {
-    fn from(value: String) -> Self {
-        Document::String(value)
-    }
-}
-
-impl From<Vec<Document>> for Document {
-    fn from(values: Vec<Document>) -> Self {
-        Document::Array(values)
-    }
-}
-
-impl From<HashMap<String, Document>> for Document {
-    fn from(values: HashMap<String, Document>) -> Self {
-        Document::Object(values)
-    }
-}
-
-impl From<u64> for Document {
-    fn from(value: u64) -> Self {
-        Document::Number(Number::PosInt(value))
-    }
-}
-
-impl From<i64> for Document {
-    fn from(value: i64) -> Self {
-        Document::Number(Number::NegInt(value))
-    }
-}
-
-impl From<i32> for Document {
-    fn from(value: i32) -> Self {
-        Document::Number(Number::NegInt(value as i64))
-    }
-}
 
 /// A number type that implements Javascript / JSON semantics, modeled on serde_json:
 /// <https://docs.serde.rs/src/serde_json/number.rs.html#20-22>
@@ -132,8 +41,6 @@ pub enum Number {
     /// 64-bit floating-point value.
     Float(f64),
 }
-
-/* ANCHOR_END: document */
 
 impl Number {
     /// Converts to an `f64` lossily.